requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
yfinance
requests
textblob
newspaper3k
nltk